requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-WTF==1.1.0
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
protobuf==3.20.1
python-dotenv==0.21.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
